requirements.txt
astroid==3.1.0
contourpy==1.2.0
cpplint==1.6.1
CTkMessagebox==2.5
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
dill==0.3.8
fonttools==4.47.2
isort==5.13.2
kiwisolver==1.4.5
matplotlib==3.8.2
mccabe==0.7.0
mpi4py==3.1.5
numpy==1.26.3
packaging==23.2
pillow==10.2.0
platformdirs==4.2.0
pybind11==2.11.1
pybind11-global==2.11.1
pylint==3.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
six==1.16.0
tenacity==8.2.3
tomli==2.0.1
tomlkit==0.12.4
typing_extensions==4.10.0
